Fix typing imports
