+++ haifa_jq/__init__.py
@@ -1,0 +1,4 @@
+"""haifa_jq package exposes jq runtime helpers."""
+from .jq_runtime import run_filter, run_filter_many, run_filter_stream, JQRuntimeError
+
+__all__ = ["run_filter", "run_filter_many", "run_filter_stream", "JQRuntimeError"]
